requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.0
transformers==4.48.2
hydra-core==1.3.2
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-ollama==0.2.3
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
owlready2==0.47